+++ NEWS.md
@@ -1,5 +1,7 @@
 # broom (development version)
 
+* `tidy.coxph()` will now pass its ellipses `...` to `summary()` internally (#1151 by `@ste-tuf`).
+
 # broom 1.0.4
 
 * Added an `intercept` argument to `tidy.aov()`, a logical indicating whether to include information on the intercept as the first row of results (#1144 by `@victor-vscn`).