+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `tidy.coxph()` will now pass its ellipses `...` to `summary()` internally (#1151 by `@ste-tuf`).
 
+* Removed speedglm as a Suggested package following the package's removal from CRAN. If the package has not made it to CRAN by the next release, the tidiers for those objects will be hard-deprecated.
+
 # broom 1.0.4
 
 * Added an `intercept` argument to `tidy.aov()`, a logical indicating whether to include information on the intercept as the first row of results (#1144 by `@victor-vscn`).