+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Removed maptools and rgeos as Suggested packages.
 
+* Addressed bug in mlogit tidiers where `augment.mlogit()` would fail if supplied a model fitted with a non-default `dfidx()`. (#1156 by `@gregmacfarlane`)
+
 # broom 1.0.4
 
 * Added an `intercept` argument to `tidy.aov()`, a logical indicating whether to include information on the intercept as the first row of results (#1144 by `@victor-vscn`).