+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Addressed bug in mlogit tidiers where `augment.mlogit()` would fail if supplied a model fitted with a non-default `dfidx()`. (#1156 by `@gregmacfarlane`)
 
+* Addressed bug in ANOVA tidiers where `tidy.anova()` would fail if passed a model with many predictors. (#1159 by `@jwilliman`)
+
 # broom 1.0.4
 
 * Added an `intercept` argument to `tidy.aov()`, a logical indicating whether to include information on the intercept as the first row of results (#1144 by `@victor-vscn`).